Added fractions and fraction logic
- Now uses fractions for timing instead of floats.

+++ src/main/java/msharp/MinecraftClasses/MinecraftFitment.java
@@ -7,16 +7,16 @@ import java.util.List;
 
 public class MinecraftFitment {
     public static NoteStructure fitToMinecraft(List<FinalNote> finalNotes){
-        float songDuration = finalNotes.get(finalNotes.size()-1).getTiming();
+        double songDuration = finalNotes.get(finalNotes.size()-1).getTiming().toDouble();
         // todo calculate timing to tick with math.round instead
-        int ticks = Math.round(songDuration*10+1);
+        long ticks = Math.round(songDuration*10+1);
 
         NoteStructure minecraftNotes = new NoteStructure();
         for(int i = 0 ; i < ticks; i++){
             ArrayList<MinecraftNote> notesThisTick = new ArrayList<>();
 
             // what notes should play this tick.
-            while(finalNotes.size() > 0 && finalNotes.get(0).getTiming() <= i*0.1+0.05){
+            while(finalNotes.size() > 0 && finalNotes.get(0).getTiming().toDouble() <= i*0.1+0.05){
                 notesThisTick.add(convertFinalNoteToMinecraftNote(finalNotes.get(0)));
                 finalNotes.remove(0);
             }
